+++ src/pages/index.tsx
@@ -1,9 +1,10 @@
-import { Route, Routes } from "react-router-dom";
+import { Navigate, Route, Routes } from "react-router-dom";
 import { Layout } from "../layout/Layout";
 import {
   Environment,
   INTERNAL_ROUTES,
   environment,
+  isStakeDisable,
   network,
 } from "../constants/constants";
 import { SwapPage } from "./swap";
@@ -12,10 +13,12 @@ import { GardenProvider } from "@gardenfi/react-hooks";
 import { useWalletClient } from "wagmi";
 import { useAccount } from "@starknet-react/core";
 import { Environment as GardenEnvironment } from "@gardenfi/utils";
+import { useSolanaWallet } from "../hooks/useSolanaWallet";
 
 function App() {
   const { data: walletClient } = useWalletClient();
   const { account: starknetWallet } = useAccount();
+  const { solanaAnchorProvider } = useSolanaWallet();
 
   return (
     <GardenProvider
@@ -30,11 +33,13 @@ function App() {
                 info: import.meta.env.VITE_INFO_URL,
                 evmRelay: import.meta.env.VITE_RELAYER_URL,
                 starknetRelay: import.meta.env.VITE_STARKNET_URL,
+                solanaRelay: import.meta.env.VITE_SOLANA_URL,
               }
             : (network as unknown as GardenEnvironment),
         wallets: {
           evm: walletClient,
           starknet: starknetWallet,
+          solana: solanaAnchorProvider ?? undefined,
         },
       }}
     >
@@ -43,9 +48,13 @@ function App() {
           {INTERNAL_ROUTES.swap.path.map((path) => (
             <Route key={path} path={path} element={<SwapPage />} />
           ))}
-          {INTERNAL_ROUTES.stake.path.map((path) => (
-            <Route key={path} path={path} element={<StakePage />} />
-          ))}
+          {isStakeDisable ? (
+            <Route path="/stake" element={<Navigate to="/" replace />} />
+          ) : (
+            INTERNAL_ROUTES.stake.path.map((path) => (
+              <Route key={path} path={path} element={<StakePage />} />
+            ))
+          )}
         </Routes>
       </Layout>
     </GardenProvider>