+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## [1.0.3]
 
+### Added
+
+- Added support for Solana.
+
 ### Fixed
 
+- Display up to 8 decimal places for balance and input amount values.
 - Optimized transaction loading to display new results without full rerender, and improved swap progress UI for refunded transactions.
 
 ## [1.0.2]